Add owned games indexing and endpoint.
