+++ index.tmpl.html
@@ -808,6 +808,11 @@
     <button type="submit">Steam Bans</button>
 </form>
 
+<form action="/owned_games" method="get">
+    <input type="text" name="steamids" placeholder="steamid|vanity name|profile url">
+    <button type="submit">Owned Games</button>
+</form>
+
 <form action="/friends" method="get">
     <input type="text" name="steamids" placeholder="steamid|vanity name|profile url">
     <button type="submit">Steam Friends</button>
