+++ jobs_sourcebans.go
@@ -3,6 +3,7 @@ package main
 import (
 	"context"
 	"log/slog"
+	"time"
 
 	"github.com/riverqueue/river"
 )
@@ -15,8 +16,9 @@ func (SourcebansArgs) Kind() string {
 
 func (SourcebansArgs) InsertOpts() river.InsertOpts {
 	return river.InsertOpts{
-		Queue:    string(QueueSourcebans),
-		Priority: int(Slow),
+		Queue:      string(QueueSourcebans),
+		Priority:   int(Slow),
+		UniqueOpts: river.UniqueOpts{ByPeriod: time.Hour * 6},
 	}
 }
 
@@ -24,6 +26,10 @@ type SourcebansWorker struct {
 	river.WorkerDefaults[SourcebansArgs]
 	database *pgStore
 	config   appConfig
+}
+
+func (w *SourcebansWorker) Timeout(_ *river.Job[SourcebansArgs]) time.Duration {
+	return time.Hour * 6
 }
 
 func (w *SourcebansWorker) Work(ctx context.Context, _ *river.Job[SourcebansArgs]) error {
