+++ limiter.go
@@ -7,35 +7,35 @@ import (
 	"golang.org/x/time/rate"
 )
 
-// type Limiter struct {
-//	waitChan chan any
-//	waitTime time.Duration
-// }
-//
-// func NewLimiter(waitTime time.Duration) Limiter {
-//	return Limiter{
-//		waitTime: waitTime,
-//		waitChan: make(chan any),
+//	type Limiter struct {
+//		waitChan chan any
+//		waitTime time.Duration
 //	}
-// }
 //
-// func (l *Limiter) start(ctx context.Context) {
-//	timer := time.NewTimer(l.waitTime)
-//
-//	for {
-//		select {
-//		case <-timer.C:
-//			l.waitChan <- true
-//			timer.Reset(l.waitTime)
-//		case <-ctx.Done():
-//			return
+//	func NewLimiter(waitTime time.Duration) Limiter {
+//		return Limiter{
+//			waitTime: waitTime,
+//			waitChan: make(chan any),
 //		}
 //	}
-// }
 //
-// func (l *Limiter) wait() {
-//	<-l.waitChan
-// }
+//	func (l *Limiter) start(ctx context.Context) {
+//		timer := time.NewTimer(l.waitTime)
+//
+//		for {
+//			select {
+//			case <-timer.C:
+//				l.waitChan <- true
+//				timer.Reset(l.waitTime)
+//			case <-ctx.Done():
+//				return
+//			}
+//		}
+//	}
+//
+//	func (l *Limiter) wait() {
+//		<-l.waitChan
+//	}
 
 type LimiterCustom struct {
 	*rate.Limiter
